requirements.txt
librosa==0.11.0
moviepy==2.2.1
numpy==2.1.3
tqdm==4.67.1
flask==3.0.0
flask-cors==4.0.0
openai-whisper>=20231117
